Add languageDropdown to login page
