+++ src/components/settings/index.test.js
@@ -3,7 +3,6 @@ import React from 'react';
 
 import { expect } from 'chai';
 import { mount } from 'enzyme';
-import sinon from 'sinon';
 
 import Settings from './index';
 import i18n from '../../i18n';
@@ -26,17 +25,7 @@ describe('Settings', () => {
     });
   });
 
-  it('renders a form and a Dropdown components', () => {
-    expect(wrapper.find('Dropdown')).to.have.length(1);
-  });
-
-  it('calls i18n.changeLanguage on chaning the value in the dropdown', () => {
-    const i18nSpy = sinon.spy(i18n, 'changeLanguage');
-
-    wrapper.find('Dropdown').simulate('click');
-    wrapper.find('Dropdown ul li').at(0).simulate('click');
-    expect(i18nSpy).to.have.been.calledWith('en');
-
-    i18nSpy.restore();
+  it('renders a LanguageDropdown component', () => {
+    expect(wrapper.find('LanguageDropdown')).to.have.length(1);
   });
 });
